chore(main): move injectHeaders to util function
